requirements.txt
requests
Pillow
bs4
beautifulsoup4
Discord.py